pyproject.toml

[tool.poetry]
name = "data-modeling-automation"
version = "0.1.0"
description = "Automated Data Modeling with AI and Data Vault 2.0"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.5.0"
python-multipart = "^0.0.6"
pandas = "^2.2.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.0"
openpyxl = "^3.1.2"
jinja2 = "^3.1.2"
pyyaml = "^6.0.1"
loguru = "^0.7.2"
python-jose = "^3.3.0"
passlib = "^1.7.4"
tenacity = "^8.2.3"
httpx = "^0.26.0"
langchain = "^0.3.0"
langchain-core = "^0.3.0"
python-dotenv = "^1.0.0"
gitpython = "^3.1.41"
redis = "^5.0.1"
fastapi-cache2 = "^0.2.1"
pydantic-settings = "^2.8.0"
langchain-groq = "^0.2.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
black = "^24.1.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.black]
line-length = 88